requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.1
numpy~=2.1.3
scikit-learn~=1.6.1
joblib==1.2.0
opencv-python-headless
Werkzeug~=3.1.3
deepface~=0.0.93
pymongo~=4.12.0
cloudinary~=1.44.0
python-dotenv~=1.1.0
pillow~=11.1.0
requests~=2.32.3